Replace old gen_rl_tasks with the scrapefile version
